requirements.txt
requests>=2.24.0
pandas>=0.25.3
Flask-JWT-Extended>=4.3.1
Flask==2.0.3
PyJWT==2.3.0
mongoengine==0.24.0
